+++ src/lib/index.ts
@@ -1,4 +1,4 @@
 export { ELogLevel, setVerbosity } from './helpers/logger';
 export type { IHeightmap, IHeightmapCoords, IHeightmapSample } from './terrain/heightmap/i-heightmap';
-export type { ILocalMapData, IVoxelMap, IVoxelMaterial } from './terrain/i-voxel-map';
 export { EPatchComputingMode, Terrain, type TerrainOptions, type TerrainStatistics } from './terrain/terrain';
+export type { ILocalMapData, IVoxelMap, IVoxelMaterial } from './terrain/voxelmap/i-voxel-map';
