+++ src/lib/terrain/terrain.ts
@@ -7,13 +7,13 @@ import * as THREE from '../three-usage';
 import { AsyncPatch } from './async-patch';
 import { HeightmapViewer } from './heightmap/heightmap-viewer';
 import { IHeightmap } from './heightmap/i-heightmap';
-import { type IVoxelMap } from './i-voxel-map';
 import { PatchFactoryCpu } from './patch/patch-factory/merged/cpu/patch-factory-cpu';
 import { PatchFactoryGpuOptimized } from './patch/patch-factory/merged/gpu/patch-factory-gpu-optimized';
 import { PatchFactoryGpuSequential } from './patch/patch-factory/merged/gpu/patch-factory-gpu-sequential';
 import { type PatchSize } from './patch/patch-factory/merged/vertex-data1-encoder';
 import { PatchFactoryBase } from './patch/patch-factory/patch-factory-base';
 import { PatchId } from './patch/patch-id';
+import { type IVoxelMap } from './voxelmap/i-voxel-map';
 import { EVoxelsDisplayMode } from './voxelmap/voxelsRenderable/voxels-renderable';
 
 type TerrainOptions = {
